test: add tests and update file structure
